Rename DataFormats directory to match required pattern
