+++ DataFormats/HGCalReco/src/classes.h
@@ -1,0 +1,5 @@
+#include <vector>
+#include <array>
+#include "DataFormats/HGCalReco/interface/Trackster.h"
+#include "DataFormats/HGCalReco/interface/TICLLayerTile.h"
+#include "DataFormats/Common/interface/Wrapper.h"
