--- DataFormats/RecoHGCal/src/classes.h
@@ -1,5 +1,0 @@
-#include <vector>
-#include <array>
-#include "DataFormats/RecoHGCal/interface/Trackster.h"
-#include "DataFormats/RecoHGCal/interface/TICLLayerTile.h"
-#include "DataFormats/Common/interface/Wrapper.h"
