+++ RecoHGCal/TICL/plugins/BuildFile.xml
@@ -6,7 +6,7 @@
 <use   name="FWCore/PluginManager"/>
 <use   name="DataFormats/HGCDigi"/>
 <use   name="DataFormats/HGCRecHit"/>
-<use   name="DataFormats/RecoHGCal"/>
+<use   name="DataFormats/HGCalReco"/>
 <use   name="CondFormats/DataRecord"/>
 <use   name="FWCore/MessageLogger"/>
 <use   name="FWCore/MessageService"/>
